move relation to common
